[Fix] Memory estimation slows down redo copy
Problem:
========
Since 8.0.30, xtrabackup introduced "PXB-2710 - Predict memory for --prepare"
to estimate memory requirements when prepare. It needs to calculate the
space of each redo record and the number of page frames.

These operations slow down redo copy. If the redo write of the
instance is heavy, the backup fails with logs like these.

"""
[Xtrabackup]  it looks like InnoDB log has wrapped around before
xtrabackup could process all records due to either log copying being too
slow, or  log files being too small
"""

Solution:
========
Since RDS does not use this function and there is no plan to use it in
the future. Introduce a new variable rds-disable-estimate-memory, which
is true, to skip these operations.