[Postfix] Fix replay rename when --lock-ddl=0
Summary
-------
Xtrabackup can not replay rename operation when --lock-ddl=0.
It might crash when in --prepare stage. The root cause is that
the rename operation cannot achieve idempotence.

And the deeper reason is that when --lock-ddl=0, it is not
guaranteed to scan the file system safely.

Fix
---
Introduce a new variable polarx_locked_for_safe_scan_fs to replace
lock-ddl. When this parameter is true, the file system will be
scanned under the protection of the backup lock to ensure safe
scanning of the file system.